requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.12

# Vector Store & Embeddings
chromadb==0.5.20
reportlab==4.4.7
sentence-transformers==3.3.1

# LLM Providers
openai==1.57.0
google-generativeai==0.8.3

# Utilities
networkx==3.4.2
numpy==1.26.4
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
